Updated analyis.py to comply to the style guide standards
I added docstrings for our methods, broke apart lines that were past the maximum length, and put the code in the format (some in  run_analysis() and some in our function) as Ryaan specified. I also made our method return a tuple of values now to print in run_analysis().