Added unit tests
Added two unit tests for our mean_rereading_time function. These tested the function on model default values, as well as tested it on the test data that was given. All passed. One note is that the rounding to 2 decimals can cause slight mistakes based on the way that Python represents floats so at times they are .01 off from what I wouldv'e expected, which is why I included rounding in the test data to mimic what the function and computer would do rather than what a human would do.