Consolidate merge conflicts and added a method to remove outliers
Resolved merge conflicts between removing outlier version and test case version, so that new version includes both. Added a method to remove outliers. Added code to mean_rereading_time_for_a_question method to store a list of the view times with outliers removed, so that the original data is not modified.